refresh trips on adding bill
the simplest route on updating the trip we're looking at seems to be to
just refresh it any time we make a change to it or its related models.

this change does that after a user submits the addBill form via a new
prop, onComplete, which can be undefined or a callback. If it exists, it
will be called on success of the api call to add a bill